core::rt: Move some TLS functions from local_sched to local_ptr
